Introduce "RequireMoreArguments" argument filter.
The reason for this parser is that having an optional argument, and a non-optional following it means that the optional argument takes effect, and then the final argument is not parsed. This allows me to specify that a first argument should only be parsed if there are a two arguments, for example.

Tests are included too.