Refactor chat handling, allow other plugins to provide tokens for Nucleus.
Plugins can use the NucleusMessageTokenService to register their own tokens. This will primarily be useful to plugins and owners that want to add their own information to the chat, but want to allow servers to control where they are placed.

Plugin tokens will be of the form {{pl:<pluginid>:<identifier>:s}}, so if plugin "bacon" offered token ID "spam", the token to use would be {{pl:bacon:spam}}. If plugin "eggs" wanted to offer the token "spam", they could do so too, their token would be {{pl:eggs:spam}}.

Nucleus' tokens have not changed.

The ":s" option is now valid on all tokens, and will add a space to the end if the token is not empty.