0.21.0-PR1
You might be wondering why I'm doing this. This is a testing build for checking API integrations with GP, and one I consider stable enough to use, but I want to ensure everything works with GP before we go for the full 0.21 release.

This will give a Maven endpoint in my repo, and allow me to make breaking changes in case something goes wrong in cross-plugin testing.