Start modularisation of data files.
While Nucleus is modular, the data files were a horrible mess of abstraction and misdirection. This system replaces the World service with modular classes. GeneralService and UserService will be converted next, this was just to make sure the system works.

Hopefully, this will make it easier to contribute, as all the code that needs to be written will be in one module, rather than trying to figure out how the data files work.

There are some utility methods to make it easy to get/set data components too.