Switch templates to the NucleusTextTemplate system.
Currently, when we build a text, we always build it from scratch. The performance of this isn't great, because there is a lot of regex involved. Thankfully, Sponge provides the handy "TextTemplate", which is really underused. So, now, when we build a string up, we can store it in here, and get the tokens dynamically, hopefully reducing the impact on the server.

No user intervention is needed, this will just work with (hopefully) improved performance.

This updates:

* /motd and /info
* /broadcast
* /afk
* Chat and /me
* Messages
* Command and Social Spy
* Connection messages
* Staff Chat

The Pagination system has a weird bug where it doesn't count empty lines when throwing out lines to the client. We now replace the empty text instances with spaces to make sure it works correctly.

Other changes:

* Updated the Default annotation on the config: don't create a full blown NucleusTextTemplate unless we don't have anything to load in.i
* Add method to join texts together and let colours/styles flow.
* Made some commands reloadable, not constructing the Config classes each time.