Overhaul kit item saving/loading
We had a previous commit that was to allow shorts to be saved/loaded. However, Configurate can be a bit frustrating sometimes, if you get an empty node from a loader, it sets that it doesn't accept the "Short" datatype. We work around this.

I also re-wrote reading from a configuration node each time, now we just need to read into a wrapper type. This should improve performance.

Fixes #685