Add chunkunload command.
I'm not sure if I'd want anything like this yet because really, memory is there to be used and continuously unloading chunks is actually not a good idea. I fear that such a command would make server owners thing this is a good thing to do all the time. It's not, but there are scenarios where unloading as many chunks as possible might be a good idea - if TPS is dropping like a stone, for example, because entities are sticking about.

That said, in that scenario, you are better off running /kill @e[!type=Player] or whatever the selector is.