Fix tab completion for the nickname argument.
As is customary in Nucleus, we resort to a slight hack. Tab completion only works if Sponge thinks the last element won't parse. When you have partial tab completion, of course, things will parse on what isn't the full argument - tab completion will either fail because it doesn't match anything, or fail because a player name will match.

The "solution" is to add data into the command context to tell the argument that we're doing a tab completion and therefore need to tab complete.

This works towards #775, fixes #788