Remove Blacklist module functionality.
The Blacklist module was poorly maintained, and there are much better solutions out there. We are there retiring the module, it does not really belong in an essentials like plugin.

A migrator for ProtectionPerms is available, GP will come soon after they version bump. An API is available to get the old data.

* Fix @Scan scanner not being restricted to a specific Java package.

Fixes #822