Add custom context key for formatting
Plugins can check for this key to suggest whether they should format the
incoming message.

Also, as the cause of "player.simulateChat" would be the SENDING player,
not the SIMULATED player due to cause/context changes, we've updated our
listeners to check the PLAYER_SIMULATED context key first. That way, we
ensure that we're formatting for the correct player.