Add placeholders/tokens.
Each of the following tokens returns text only if the status is active.

{{afk}}
{{jailed}}
{{muted}}
{{flying}}
{{frozen}}

The text that is returned can be customised. For afk, if you have the token {{afk|yes}}, the token will return "yes" if the player is AFK, nothing otherwise. If you have the token {{afk|yes|no}}, then "yes" will be returned if true, "no" otherwise. The text, of course, can be customised with standard colour codes.

{{jail}} - returns the jail a player is in, if any

Fixes #923