Refactor some group template logic to allow some servers to reduce overhead
There is now a new permission option, "nucleus.chat.group" that can be
applied to groups or users to force the use of that group when selecting
a group template. This bypasses potentially expensive checks during
runtime by not having to calculate the parent groups for a player.

Servers are then encouraged to turn off the new chat config option
"chat.templates.check-permission-groups" once they have added the
required options to the groups. This should result in a performance gain
by servers.

This will also stop rare watchdog errors killing the server when players
chat - the time the system is allowed to take to perform these
operations is 100ms, which is enough to slow a server down to a crawl
but would be a more noticable problem. Nucleus will spit the error into
the console, allowing for servers to be shut down cleanly.

Works around #1073